Do not force users to specify '.so' or '.dll' manually for linkshared=1.
This is legacy behaviour back when it was hard to do. These days, the toolchain can figure out an appropriate name for binaries and shared libraries, but only the binary code path was updated.

TESTED=Added one tiny test to make sure it is no longer an error to create such a cc_binary. I don't know if it's safe to assert on exact name, seeing as the tests would then be platform specific.
RELNOTES: linkshared=1 in cc_binary no longer requires '.so' or '.dll' in the target name
PiperOrigin-RevId: 313398118