Allow "undoing" the --wait_fix flag flip.
Passing -W (or --wait_fix) to the process wrapper enables a fix that has
caused quite a bit of trouble in the past.  Having the flag is good for
a controlled rollout, but we should also have a mechanism to undo the
effects of the rollout (so that when we flip the default in a Bazel
release users still have a way out).

To accomplish this, make --wait_fix take an optional argument that, if
is "no", restores the old (broken) behavior.

In preparation to address #10245.

RELNOTES: None.
PiperOrigin-RevId: 313577008