Introduce AggregatedSpawnMetrics
This allows aggregating `SpawnMetrics` per different execution modes (remote,
worker, local). So now we'll be able to differentiate between the aggregated
metrics for remote, local and worker executions, which is useful because we
probably don't want to aggregate the queuing time for remote and worker
execution together.

This change doesn't yet change any output, but in the future we should be able
to report the metrics not only for remote execution, but also for local actions
and workers.

The next step is populating more metrics (not only execution time) for local
and worker execution.

RELNOTES: None.
PiperOrigin-RevId: 314097914