Persistent tests: fix NPE for most test rules
Only the java_test rule currently supports the persistent test runner, and
all other test rules throw NPE. Instead, automatically fall back to the
non-persistent test runner for rules that don't support it, avoiding the NPE.

This requires updating the infrastructure to *not* just check the test config,
but also whether an individual test *action* has the persistent runner enabled.

This allows enabling this feature by default, but having it only apply to rules
that support it.

Note that I am not sure if the persistent test runner works in Bazel since
there are no public tests for it.

Fixes #11519.

Change-Id: I321e7497267e77e668209a541600908b38a32464