Add option to preserve input tar mtimes in TarFileWriter
Also backfill unit tests of default mtime functionality.

This is the first in a short sequence of changes aimed at fixing static file caching issues in applications deployed using the container_image rule in [rules_docker](https://github.com/bazelbuild/rules_docker). That rule relies on TarFileWriter, which currently copies files from one tar to another and overwrites the input files' mtimes, which causes the Last-Modified HTTP header to always return the same value and breaks browser caches.

This PR contains an additive change in the form of an optional parameter to TarFileWriter that copies mtimes from the input tar files to the output tar. It should not affect any existing behavior.

Closes #11028.

PiperOrigin-RevId: 314323849