Clean up artifact retrieval methods in ActionExecutionValue.
All callers of getArtifactValue were doing a checkNotNull on the return value, which is exactly what getExistingFileArtifactValue does. Migrate them so that we only need one retrieval method. Loosen the parameter to Artifact and do the instanceof check in one place, so any bugs get an appropriate error message.

Additionally, use the recently added isChildOfDeclaredDirectory method to determine which map to search for the artifact, instead of unnecessarily checking both.

RELNOTES: None.
PiperOrigin-RevId: 314380367