URL encode "%" due to rules_jvm_external paths
Fixes a problem where the manifest classpath is incorrectly written due to not URL-encoding symbols.

The paths provided in the manifest when the command line is too long is supposed to be a URL, which means we have to URL-encode special characters. `rules_jvm_external` generates a bunch of artifacts with paths that are themselves a URL-encoded version of their maven server host so we have a bunch of "%40" strings in the path due to usage of "@" for user@host http auth.