C++: Fix crash with repeated precompiled library in srcs
With a precompiled library repeated in the srcs of a cc_library we got a crash
when trying to add them both to an immutable map because they had the same
identifier. Since it's probably undesired behavior to link the same library
twice built in different configurations which is what happened as originally
reported in b/156775573, the behavior will still be an error, a proper error
though instead of a crash.

RELNOTES:none
PiperOrigin-RevId: 314680827