Make process-wrapper exit as soon as a timed-out process handles SIGTERM.
Before this change, passing --timeout=X --kill_delay=Y would cause the
process-wrapper to send a SIGTERM to the subprocess after X seconds... and
then unnecessarily wait for Y seconds to send a SIGKILL.  If the process
exited after X passed but before Y expired, we wouldn't notice.

The reason is that we were using kill(pid, 0) to check if the process had
already vanished after a SIGTERM, expecting this call to return an error
once the process is gone. The problem is: the process is never gone until
a wait() is done, as otherwise it remains as a zombie, and sending signals
to a zombie process doesn't return an error.

Given that we cannot wait in the signal handler, make the signal handler
install a second timer to send the SIGKILL once the Y seconds pass, and
then let the main logic in the process-wrapper await for the subprocess'
termination and finalization of the whole process tree.

Part of #7818.

RELNOTES: None.
PiperOrigin-RevId: 314905347