Handle some interrupts better during local process execution.
InterruptedIOException is an IOException, but it represents an interrupt.
If we get this, we should correctly propagate it as an InterruptedException
so as not to confuse the dynamic scheduler's branch cancellation.

Furthermore, there are some cases where we spawn a process, terminate it,
and don't notice that there was an interrupt.  Trying to report the process'
exit status in that case would be a mistake so handle that as well.  The
dynamic scheduler should already be resilient to this by checking interrupted
statuses on its own, but better be explicit.

I've observed these experimentally during some very large builds using the
new dynamic scheduler and --experimental_local_lockfree_output but haven't
been able to come up with simple test cases to show them.

Part of #7818.

RELNOTES: None.
PiperOrigin-RevId: 314935379