Move management of .bzl builtin symbols to StarlarkBuiltinsFunction
This change centralizes knowledge of the set of predeclared symbols for .bzl files (and in a future change, perhaps BUILD/WORKSPACE files) in StarlarkBuiltinsFunction. For WORKSPACE- and @builtins-loaded .bzls, where the set of predeclareds does not change, it is computed once at Skyframe initialization time. For BUILD-loaded .bzls, the original set of predeclareds is also computed and saved at init time, but in the future we'll request the injected symbols from exports.bzl and save a modified environment in the returned StarlarkBuiltinsValue.

ConfiguredRuleClassProvider and PackageFactory now provide a way to distinguish "rule logic" symbols (e.g. `CcInfo`) from generic symbols (e.g., `rule()`). This will be used in a future CL to prohibit injecting new values for generic symbols.

Work toward #11437.

Other changes:

StarlarkBuiltinsValue:
- Store the desired environment, rather than the delta between the original environment and the desired one. This makes it so we won't have to reapply the delta (and validation) on every load.

BzlLoadFunction:
- Move Module creation out of executeModule(), and rename the latter to executeBzlFile().
- Simplify comment at call to executeBzlFile(), pass in a listener instead of whole Environment.

StarlarkBuiltinsFunctionTest:
- Use mock symbols in the test RuleClassProvider. Update tests to use mock symbols instead of non-existent names, since those will break going forward.
- Add tests of getNativeRuleLogicBindings(). (This seemed a better home for them than PackageFactoryTest, particularly because there's no generic RuleClassProviderTest.)
- Add tests of exports.bzl files that perform loads.

AbstractPackageLoader:
- Let it know about StarlarkBuiltinsFunction, apparently it's needed for tests (in a follow-up CL). Added a comment to SkyframeExecutor.

WorkspaceFactory:
- inline a helper method

RELNOTES: None
PiperOrigin-RevId: 315101101