Prevent "Deleting stale sandbox" messages when things go right.
Make Bazel clean up the sandbox base's top directory after the end of each
build and during server shutdown to prevent this directory from staying
behind.

This directory is only supposed to stay behind when --sandbox_debug is
used, and the startup message is intended to tell the user that a heavy
operation is about to happen -- but we should not print it unnecessarily.

This was the original intent when this message was added but for some
reason I missed this obvious case.  Added a test now.

Fixes #8525.

RELNOTES: None.
PiperOrigin-RevId: 315247251