Expand error message for LTO compilation with list of missing files in imports.
LtoBackendAction checks whether bitcode files are available for all of the
imports. Expand the error message not only to indicate there was a difference
but also to list the missing files. Please note that there cannot be extra
files given the constructed NetsedSet is always a subset of the imports.

PiperOrigin-RevId: 315269362