Improve error handling for builtins injection
BuiltinsFailedException gets wrapped as a BzlLoadFailedException in BzlLoadFunction, and vice versa in StarlarkBuiltinsFunction. This makes for more verbose errors, but generally users who are not Bazel developers shouldn't see these errors.

The tests assert against the long form of the error rather than a substring to demonstrate the message in context.

Also slightly refactored the way errors in loaded bzls are reported, for uniformity between inlining and non-inlining code paths.

Work toward #11437.

RELNOTES: None
PiperOrigin-RevId: 315303094