Factor BzlLoadFunction.InliningState into callers of the inlining code path
InliningState becomes package-visible and gains some documentation. It is now an opaque object that replaces the set of visited bzl keys for a BUILD file.

This prepares the way for a follow-up CL that has StarlarkBuiltinsFunction call BzlLoadFunction's inlining code path with an existing InliningState (obtained from prior BzlLoadFunction evaluation). It allows us to have it call computeInline(), the same entry point which PackageFunction calls, rather than exposing computeInlineWithState() to package visibility.

computeInlineWithState() is renamed to computeInlineCachedData(), since its distinguishing feature is now only that it returns the cached data wrapper, and not that it accepts InliningState.

Work toward #11437.

RELNOTES: None
PiperOrigin-RevId: 315309686