Add inlining code path to StarlarkBuiltinsFunction, and tests
StarlarkBuiltinsFunction grows a computeInline() method that accepts the calling BzlLoadFunction's InliningState and forwards it back to BzlLoadFunction#computeInline. It follows the same RecordingSkyFunctionEnvironment unwrapping convention as BzlLoadFunction uses to recursively call itself inline.

CachedBzlLoadData had some assertions on how its API could be used. These appear to be overly specific and are removed. (This allows the transitive dep on builtins to be resolved in any order relative to other deps.)

The tests are just clones of the existing tests with inlining turned on.

Work toward #11437.

RELNOTES: None
PiperOrigin-RevId: 315338126