Add the --experimental_local_retries_on_crash option.
This is intended to be used with --experimental_local_lockfree_output,
which causes lots more local process to be eagerly terminated, and is
an attempt to workaround an odd OSXFUSE bug that introduces data corruption
on reads for which we have no diagnostic nor fix yet.

Part of #7818.

RELNOTES: None.
PiperOrigin-RevId: 315485365