bazel NestedSet: eliminate size subfield by recording it in memo
This liberates 30 bits of space into which we can record the depth
(in a follow-up).

Janak had suggested getting rid of the size altogether, but this
is no small task because of the 6 existing uses; see discussion in
commit ed9d6390eb78303631c25b3d4776cb07eeded544. The size is also used internally by toList to allocate
an array of the correct size.

Rather than block my progress on that project, this change encodes
the size in the final bytes of the memo array, using a reverse
varint encoding. This would appear to add 1-3 three bytes to each
flattened set, but in practice is almost free as there is usually
slack in the array.

The Usual Benchmark shows a RAM increase of +0.01%.

See saga: commit 841ac61469ed61325acbcb134b6d372316f46540, commit ed9d6390eb78303631c25b3d4776cb07eeded544, commit 867d232f27aa762d4b15e811df18b1939cc1a34b.

PiperOrigin-RevId: 315981081