Support link stamping in `apple_binary` rules
Add a `stamp` attribute to the `apple_binary` rule, and give it the same
semantics as `stamp` in `cc_binary`: When building a binary with
`stamp = 1`, Bazel will build all code from all `linkstamp` attributes
on all the binary’s transitive dependencies and include that code in the
final link. This lets you embed information about the build into the
final link without busting cache.

RELNOTES:
`apple_binary` rules now accept the `stamp` attribute with the same
semantics that it has in `cc_binary` rules.
PiperOrigin-RevId: 316519143