Automated rollback of commit 72bd8709dd71bd20f65938e926a1b41ddb3f5bec.
*** Reason for rollback ***

Android Studio didn't end up needing this

*** Original change description ***

Enable GenClass to take Java source jars as inputs to extract their
corresponding class files from an input class jar.

No parsing of the Java files is done, so the classes to extract are just based
on the Java file's path in the source jar (i.e. the only compilation unit is
assumed to be a class that matches the name of the file, and that the Java
file's package matches its path in the jar.)

RELNOTES: None.
PiperOrigin-RevId: 316563736