Add option rename dll on Windows
@meteorcloudy
This PR will add an option rename dll on windows.
It won't change any name in the cc_binary and cc_library. It will only change the name when we copy the DLL to the binary directory and the name in the def file
https://github.com/bazelbuild/bazel/blob/1c03c8c4fac1e1028fcb5bb342da3a7fdfc88327/src/main/java/com/google/devtools/build/lib/rules/cpp/CcBinary.java#L1054

It is just a draft and it can not handle correctly the file name extension ( only dll for now )

Closes #11524.

PiperOrigin-RevId: 316650491