Do not replace .md with .html in non-relative links in docs.
Until now, we had docs pointing to non-existent .html files in github and even
wrongly replaced README.md with README.html in some code examples.

For example, in configurable-attributes doc, we had:

<a href="https://github.com/bazelbuild/bazel-skylib/blob/master/docs/selects_doc.html"><code class="highlighter-rouge">selects</code></a>

After the fix, we will have:

<a href="https://github.com/bazelbuild/bazel-skylib/blob/master/docs/selects_doc.md"><code class="highlighter-rouge">selects</code></a>

Full recursive diff of `./scripts/serve-docs.sh --target` output
with the result of this fix:
https://gist.github.com/tetromino/fa590eff74db10ac0815773cb46d821b

Note that parsing markdown or html with sed is generally impossible to
get right, but since we're already doing it, we may for now continue; the
alternative would be to add an intelligent markdown and html parser to
the pipeline.

Fixes https://github.com/bazelbuild/bazel/issues/6285

RELNOTES: None.
PiperOrigin-RevId: 316705091