Introduce `pythonShebang` variable for nix packages collection
Building on Nix / NixOS we need to avoid using `/usr/bin/env` to determine the python interpreter (or other executables).

This change makes it easier for us to replace the shebang line when building the derivation of bazel with a fixed path to a python3 executable.

See NixOS/nixpkgs#87726 for details.

Closes #11535.

PiperOrigin-RevId: 316886082