Only add a file path into the action cache if it's non-mandatory.
For actions that discover inputs, we store every input's execPath in the action cache, so that the corresponding artifact can be retrieved later when we load the cache entry. However, there's no need to do this for mandatory inputs (and outputs) since we already have their execPaths from the action itself.

The benchmark results showed a tiny improvement in retained heap size (~0.20%).

PiperOrigin-RevId: 317301234