Shorten the hash suffix in the output path caused by a config transition.
We make it 48-bit, which should be sufficient to avoid hash collisions for the tiny handful of output paths we generate. In return, we save 52 characters in the output path. This is important for Windows with its tiny MAX_PATH, but should lead to more ergonomic output paths in general.

RELNOTES:
PiperOrigin-RevId: 318072937