Revert the Starlark debugger flags to their old name.
IDEs are using those flags, we don't want to break them now. We could have set
oldName so that both names would be supported. However, we plan to make the flags
non experimental in the future. It's not worth doing two renamings, and we can
live with the old names for now.

RELNOTES: None.
PiperOrigin-RevId: 318534237