Use the Maven artifact ID as the canonical_id in jvm_maven_import_external.
This affects how downloaded Maven artifacts are looked up from the cache. Before this, cache lookups would only be based on the hash. So changing the artifact ID (like bumping the version) but forgetting to change the hash would result in the cache lookup silently fetching the old version (with the matching hash).

Now, the cache lookup will use the artifact ID and the hash, so bumping the artifact ID but forgetting to update the hash will result in a cache miss, a redownload of the artifact, and an error when the hash doesn't match.

This canonical_id parameter was added for exactly this purpose in https://github.com/bazelbuild/bazel/commit/c917ab208942e7b250e4dc928863a56ea2bdaee3 and https://github.com/bazelbuild/bazel/issues/5144

Fixes https://github.com/bazelbuild/bazel/issues/10353

PiperOrigin-RevId: 318830199