Configure coverage and runfiles for sh_library
InstrumentedFilesProvider is already configured for sh_binary and sh_test, it's confusing for this to not be configured the same way. This means that you might get coverage for binaries depended on via a data dependency, even if that dependency is indirect (e.g. an sh_library rule whose sources run the binary provides the data dependency).

sh_library should also gather runfiles from transitive dependencies in the same way as sh_test and sh_binary.

RELNOTES: Configure coverage and runfiles for sh_library.
PiperOrigin-RevId: 319092861